+++ src/Products/MeetingCharleroi/setuphandlers.py
@@ -10,6 +10,8 @@
 #
 import datetime
 from datetime import timedelta
+
+from Products.MeetingCommunes.config import PORTAL_CATEGORIES
 from collective.eeafaceted.dashboard.utils import addFacetedCriteria
 from imio.helpers.catalog import addOrUpdateIndexes
 from plone import api
@@ -18,7 +20,7 @@ from plone.dexterity.utils import createContentInContainer
 from imio.helpers.content import get_vocab
 from plone.memoize.forever import _memos
 from Products.CMFPlone.utils import _createObjectByType
-from Products.MeetingCharleroi.config import CC_ARRET_OJ_CAT_ID
+from Products.MeetingCharleroi.config import CC_ARRET_OJ_CAT_ID, COUNCIL_DEFAULT_CLASSIFIER
 from Products.MeetingCharleroi.config import COMMUNICATION_CAT_ID
 from Products.MeetingCharleroi.config import COUNCIL_DEFAULT_CATEGORY
 from Products.MeetingCharleroi.config import COUNCIL_DEFAULT_CLASSIFIER